Functions moved from xscmanager.go to errorreport.go to match the other xsc services files
