+++ utils/xsc/xscmanager.go
@@ -1,16 +1,10 @@
 package xsc
 
 import (
-	"fmt"
-
 	"github.com/jfrog/jfrog-cli-core/v2/utils/config"
 	"github.com/jfrog/jfrog-cli-core/v2/utils/coreutils"
-	clientutils "github.com/jfrog/jfrog-client-go/utils"
-	"github.com/jfrog/jfrog-client-go/utils/log"
-	"github.com/jfrog/jfrog-client-go/xsc"
-	"github.com/jfrog/jfrog-client-go/xsc/services"
-
 	clientconfig "github.com/jfrog/jfrog-client-go/config"
+	"github.com/jfrog/jfrog-client-go/xsc"
 )
 
 const minXscVersionForErrorReport = "1.7.7"
@@ -33,29 +27,4 @@ func CreateXscServiceManager(serviceDetails *config.ServerDetails) (*xsc.XscServ
 		return nil, err
 	}
 	return xsc.New(serviceConfig)
-}
-
-func SendXscLogMessageIfEnabled(errorLog *services.ExternalErrorLog, xscManager *xsc.XscServicesManager) error {
-	if !IsReportLogErrorEventPossible(xscManager) {
-		return nil
-	}
-	return xscManager.SendXscLogErrorRequest(errorLog)
-}
-
-// Determines if reporting the error is feasible.
-func IsReportLogErrorEventPossible(xscManager *xsc.XscServicesManager) bool {
-	xscVersion, err := xscManager.GetVersion()
-	if err != nil {
-		log.Debug(fmt.Sprintf("failed to check availability of Xsc service:%s\nReporting to JFrog analytics is skipped...", err.Error()))
-		return false
-	}
-	if xscVersion == "" {
-		log.Debug("Xsc service is not available. Reporting to JFrog analytics is skipped...")
-		return false
-	}
-	if err = clientutils.ValidateMinimumVersion(clientutils.Xsc, xscVersion, minXscVersionForErrorReport); err != nil {
-		log.Debug(err.Error())
-		return false
-	}
-	return true
 }
